* Apps/EOModelEditor/AdaptorsPanel.m         * Apps/EOModelEditor/Preferences.m         * EOAccess/EOAttribute.m         * EOAccess/EOExpressionArray.m         * EOAccess/EORelationship.m         * EOAdaptors/PostgreSQLAdaptor/LoginPanel/PostgreSQLLoginPanel.m         * EOAdaptors/SQLiteAdaptor/SQLite3Channel.m         * EOControl/EOCheapArray.m         * EOControl/EOMutableKnownKeyDictionary.m         * EOControl/EOPrivate.h         * EOControl/EOPrivate.m         * EOControl/Makefile.preamble         * GDL2Palette/ConnectionInspector.m         * GDL2Palette/DisplayGroupInspector.m           Fix warnings about conflicting return types or paramters
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gdl2/trunk@36241 72102866-910b-0410-8b05-ffd578937521

+++ Apps/EOModelEditor/AdaptorsPanel.m
@@ -153,15 +153,15 @@ static NSArray *_adaptorNames;
 
 /* NSBrowser delegate stuff */
 
-- (int) browser:(id)sender numberOfRowsInColumn:(int)column
+- (NSInteger) browser:(id)sender numberOfRowsInColumn:(NSInteger)column
 {
   return [_adaptorNames count];
 }
 
 - (void)browser:(NSBrowser*)sender
         willDisplayCell:(NSBrowserCell*)cell
-                  atRow:(int)row
-                 column:(int)column
+                  atRow:(NSInteger)row
+                 column:(NSInteger)column
 {
   [cell setLeaf:YES];
   [cell setTitle: [_adaptorNames objectAtIndex:row]];
